Adds Makefile to run commands.
Poetry doesn't support scripts like NPM does, so make it's the
alternative. In fact, they use it themselves. It works fine, except for
the whole ARGS thing to pass extra arguments. But it is what it is.

The `run` comment in the Makefile runs the `run.py` script. To preserve
the absolute import in that script, I'm choosing to run it as a module
(`python -m incubator.run`) instead of running the script directly.