Adds model with GloVe.
I decided to keep the different versions of each model in the same file.
The class receives the layers that can change in the constructor, and
the functions can be used to directly build the alternatives. This is to
replace the huge file with the builders that I had in another project.