Adds CLI interface for training
It's on the run.py file, with the subcommand 'train'. Other executables
will be ran as subcommands of run.py. I chose to do that to leave
train.py as clean as possible, and moved all logic to loading a dataset
and model in the run.py, as I'm likely to reuse some of it for eval.