Adds option to unfreeze GloVe weights. Fixes bug in acc reporting
The bug: the reported accuracy during train was wrong because I using
the wrong length.

Freeze: by default `Embeddings.from_pretained` sets the weights'
required_grad to False, but gives the option to override this (the
freeze option, which defaults to True). I added to option to change this
to the Glove Simple Classifier initialiser.