Adds saving and loading of processed Glove file
I save it/load it in the `load_glove` function. The path is derived
from the model path, which makes sense since it depends on the dataset
parser that the model uses. It will create duplicates in the future, but
it's good enough for now.

I didn't implement the same for the dataset because its loading is
reasonable for now (0.5s for the dev split).

Also fixed the tests. The test for training broke because of WandB. I
mocked its `log` function out, as it was the only one being used. I also
had to suppress some deprecation warnings as WandB uses a _very_
outdated version of the `gql` library. I tried to do that in the
pyproject.toml file, but pytest wouldn't accept the ignore coming from
there. Also tried from the patch inside the test file, no cigar either.