Adds LinearCnn model, tests for new models and renames some files
I decided to rename the files because every model will be a classifier,
so it makes no sense to make everyone have it in their names.

Added tests for LinearRnn and LinearCnn models. For Cnn I have two: one
for normal behaviour, and one for the case when the input sentence
doesn't have enough characters for the filter to work. This is expected
to fail, and the test assures so.

In reality we just have to be careful
that we never have a batch with sentences smaller (as in, number of
tokens) than the largest filter.