Fixes bug on Path/str handling in MeldLinearDataset constructor
It was accepting only Path and DataFrame objects. When I forgot and
passed a str object, it didn't trigger the Path test and tried to open
as a DataFrame. I decided to add str to the possible types, instead of
converting it from Path everywhere.

Maybe I should've enforced the use of Path everywhere that is supposed
to be a Path... But supporting str is still valuable because it might be
a common bug. I don't see the point on _forcing_ Path usage. I only
intend to use Path instead of os.path for actual path manipulation. For
direct file opening I don't see a point.