Separates the Vocabulary from the dataset
Not really separates, more like gives the option to pass it as a
parameter. The idea is that the train dataset would build the Vocabulary
on initialisation and then the dev dataset would use take it and use it
instead of build their own.

I still have to figure out how I'm going to do evaluation later. I'll
have to persist the Vocabulary, and it shouldn't be hard.