Fixes bug with Contextual dataloader.
I'm not entirely sure why these errors didn't appear with the Linear
dataloader, but they came up now. There were two types:

1. Invalid indexes
2. Invalid rows

For (1), it was a case of the training data not containing any
utterances for dialogue #60. It was supposed to be there (there are
~1000 dialogues), but that index was skipped for some reason.

(2) happened because of my pre-processing. I remove punctuation, and for
some reason there are utterances which are just punctuation (for
example, there's a "!" from Joey on item train/369). This leads to an
utterance without tokens. With batch size = 1, I just have an empty
tensor.

Because of this I had to come up a way to _skip_ an item. There was an
'easy' approach of returning None from the dataset and filter None's
at the `collate_fn`. I didn't like this because it meant I could never
test models with batch size = 1, since an invalid item would mean an
empty batch, and that can't work.

What I ended up doing was replicating the solution from the `nonechunk`
library[^1]. I move the original `__getitem__` to a separate function
(here it's called `_item`) and make it return None if the item is
invalid. The new `__getitem__` loops starting from the given index until
it finds a non-None item. This solves the other problems but means that
I will have duplicate items during training. I am fine with that
trade-off.

[^1]: https://github.com/msamogh/nonechucks/blob/01f29730de29f5f0e62821650fe21cb1f63ad4a9/nonechucks/dataset.py#L87