Adds ContextualRnn model
The main issue is that the nn.LSTM _only_ accepts 3-D (batch,
seq_len, emb_dim) input (batch-first here). This meant I had to reshape
the 4-D tensors to make it work. Wasn't a problem, just a surprise, as I
used the AllenNLP wrapper before and it worked directly there.

Other than that, it was easy.