Fixes errors and warnings
Windows defaults to latin charset, so there was some trouble reading
Glove's files. I explicitly added the UTF-8 encoding to the open call.
Maybe I should start doing that everywhere.

There was also a deprecation warning on a use of Tensor::nonzero. It
looks like this warning is actually a bug; the overload I was using is
fine, but was being recognised as something else. I explictly set the
parameter to make sure I was using the right overload, and the warning
stopped.